refactor: use shared QuickAdd component
