+++ src/app/page.tsx
@@ -1,6 +1,8 @@
 "use client";
 import { insertTask } from "@/lib/data";
 import React, { useEffect, useMemo, useState, useCallback } from "react";
+import QuickAdd from "@/components/QuickAdd";
+import { FixedSizeList as List } from "react-window";
 import { loadProjects, loadTasks, saveProjects, saveTasks } from "../lib/storage";
 
 /* ===========================
@@ -43,6 +45,7 @@ function Card(p: React.HTMLAttributes<HTMLDivElement>) {
   const { className = "", ...rest } = p;
   return <div {...rest} className={`${TOK.card} p-4 ${className}`} />;
 }
+const TBody = React.forwardRef<any, any>((props, ref) => <tbody ref={ref} {...props} />);
 function Logo({ className = "" }) {
   return (
     <svg viewBox="0 0 128 128" className={className} xmlns="http://www.w3.org/2000/svg" aria-label="Flowgrid logo" role="img">
@@ -643,7 +646,7 @@ function ProjectPage({ project, tasks, onBack, onSaveProject, onAddTask, onOpenT
       <Card className="mt-4">
         <div className="flex items-center justify-between">
           <div className="font-semibold">Quick Add</div>
-          <QuickAdd onAdd={onAddTask} />
+          <QuickAdd projects={[project]} defaultProjectId={project.id} onAdd={(title) => onAddTask(title)} />
         </div>
       </Card>
     </div>
@@ -1015,18 +1018,6 @@ function Dashboard({ tasks, projects, onCardFilter }:{ tasks:Task[]; projects:Pr
     </>
   );
 }
-
-/* ---------- Quick Add ---------- */
-function QuickAdd({ onAdd }:{ onAdd:(title:string)=>void }){
-  const [title,setTitle]=useState("");
-  return (
-    <div className="flex items-center gap-2">
-      <Input placeholder="Quick task" value={title} onChange={e=> setTitle(e.target.value)} onKeyDown={e=> e.key==="Enter" && title.trim() && (onAdd(title.trim()), setTitle(""))} />
-      <Button onClick={()=>{ if(title.trim()){ onAdd(title.trim()); setTitle(""); } }}>Add</Button>
-    </div>
-  );
-}
-
 /* ---------- Footer ---------- */
 function Footer(){
   return <footer className="border-t border-neutral-200 bg-white text-neutral-500 text-xs text-center py-4">Created by — Mohd. Izhan</footer>;
@@ -1049,6 +1040,7 @@ export default function Page(){
   const [showTask,setShowTask]=useState<Task|null>(null);
   const [showNewProject,setShowNewProject]=useState(false);
   const [paletteOpen,setPaletteOpen]=useState(false);
+  const activeProject = useMemo(() => projects.find(p => p.id === activeProjectId) || null, [projects, activeProjectId]);
 
   const notifications = useMemo(()=>{
     const now=new Date(); const list:string[]=[];
@@ -1129,16 +1121,6 @@ const saveTask = (updated: Task)=> setTasks(prev=> prev.map(t=> t.id===updated.i
   const createProject = (p:Project)=> setProjects(prev=> [p,...prev]);
   const updateProject = (p:Project)=> setProjects(prev=> prev.map(x=> x.id===p.id? p: x));
 
-  const quickAddTask = (title:string, projectId?:string)=>{
-    const id=Math.random().toString(36).slice(2);
-    const newTask:Task={ id, projectId, title, status:"Backlog", priority:"Medium", assignees:["Me"], tags:[], description:"", checklist:[], createdAt:nowIso(), updatedAt:nowIso() } as Task;
-    setTasks(prev=> [newTask,...prev]);
-  };
-
-  const createProject = (p:Project)=>{ setProjects(prev=> [p,...prev]); };
-  const updateProject = (p:Project)=>{ setProjects(prev=> prev.map(x=> x.id===p.id? p: x)); };
-  const saveTask = (t:Task)=>{ setTasks(prev=> prev.map(x=> x.id===t.id? t: x)); };
-  const deleteTask = (id:string)=>{ setTasks(prev=> prev.filter(t=> t.id!==id)); };
   const runCommand = (cmd:string,arg?:string)=>{
     if(cmd==="new-task"){ const title=prompt("Task title"); if(title) quickAddTask(title); }
     if(cmd==="go-dashboard") setPage("dashboard");
@@ -1157,7 +1139,7 @@ const saveTask = (updated: Task)=> setTasks(prev=> prev.map(t=> t.id===updated.i
           <div className="flex items-center justify-between">
             <div className="flex items-center gap-2"><Input placeholder="Search tasks… (⌘K for commands)" value={query} onChange={e=> setQuery(e.target.value)} /></div>
             <div className="flex items-center gap-2">
-              <QuickAdd onAdd={(title)=> quickAddTask(title, activeProjectId || undefined)} />
+              <QuickAdd projects={projects} defaultProjectId={activeProjectId} onAdd={quickAddTask} />
             </div>
           </div>
 
@@ -1188,60 +1170,5 @@ const saveTask = (updated: Task)=> setTasks(prev=> prev.map(t=> t.id===updated.i
       {showTask && <TaskModal task={showTask} projects={projects} onSave={(t)=> { saveTask(t); setShowTask(null); }} onDelete={(id)=> deleteTask(id)} onClose={()=> setShowTask(null)} />}
       <CommandPalette open={paletteOpen} onClose={()=> setPaletteOpen(false)} onRun={runCommand} projects={projects} />
     </main>
-  );
-}
-
-
-/* ---------- Quick Add (with project picker) ---------- */
-function QuickAddWithProject({
-  projects,
-  defaultProjectId,
-  onAdd,
-}: {
-  projects: Project[];
-  defaultProjectId?: string | null;
-  onAdd: (title: string, projectId?: string) => void;
-}) {
-  const [title, setTitle] = useState("");
-  const [projectId, setProjectId] = useState<string>(defaultProjectId || "");
-  useEffect(() => { setProjectId(defaultProjectId || ""); }, [defaultProjectId]);
-
-  const go = () => {
-    const t = title.trim();
-    if (!t) return;
-    onAdd(t, projectId || undefined);
-    setTitle("");
-  };
-
-  return (
-    <div className="flex items-center gap-2 bg-white border border-neutral-200 rounded-xl px-2 py-2 shadow-sm">
-      <input
-        className="flex-1 px-3 py-2 rounded-lg outline-none"
-        placeholder="Quick task…"
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-        onKeyDown={(e) => e.key === "Enter" && go()}
-      />
-      <select
-        className="px-3 py-2 rounded-lg border border-neutral-200 bg-white"
-        value={projectId}
-        onChange={(e) => setProjectId(e.target.value)}
-        title="Select project"
-      >
-        <option value="">No Project</option>
-        {projects.map((p) => (
-          <option key={p.id} value={p.id}>
-            {p.icon || "📁"} {p.name}
-          </option>
-        ))}
-      </select>
-      <button
-        onClick={go}
-        disabled={!title.trim()}
-        className="px-3 py-2 rounded-xl bg-black text-white disabled:opacity-50 hover:opacity-90 active:scale-[.99]"
-      >
-        Add
-      </button>
-    </div>
   );
 }
