+++ tests/ProjectPage.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
-import { ProjectPage } from '../src/app/page';
+import ProjectPage from '../src/components/ProjectPage';
 
 const project = { id: 'p1', name: 'Proj', status: 'Active', tags: [] };
 const task = {
